libfm: fix a build issue
Avoids this build issue:
| rmdir: failed to remove `/srv/home/nitin/builds/build-gcc47/tmp/work/i586-poky-linux/libfm-0.1.17-r4/image/usr/lib/gio/modules/': No such file or directory
NOTE: package libfm-0.1.17-r4: task do_install: Failed

+++ meta/recipes-support/libfm/libfm_0.1.17.bb
@@ -22,8 +22,12 @@ SRC_URI[sha256sum] = "1740681cff4cd4c5a2eaa9805d8898269cfb6a49a0bda0acb242def15b
 inherit autotools pkgconfig
 
 do_install_append () {
-	rmdir ${D}${libdir}/gio/modules/
-	rmdir ${D}${libdir}/gio/
+	if [ -d  ${D}${libdir}/gio/modules/ ] ; then
+		rmdir ${D}${libdir}/gio/modules/
+	fi
+	if [ -d  ${D}${libdir}/gio ] ; then
+		rmdir ${D}${libdir}/gio/
+	fi
 }
 
 PACKAGES += "${PN}-mime"
